refactor(ts): dissolve simulation/adapters/ layer
Relocate the two adapter functions to their natural homes and delete the
adapters/ directory (an ugly merge-era layer):
- compileEndaxisScenario -> simulation/compileEndaxisScenario.ts (Endaxis compile
  entry point wrapping compiler/compileScenario)
- projectOptimizerResult -> simulation/projection/projectOptimizerResult.ts
  (alongside the other project* modules)

The simulation.ts composable already orchestrates these two standalone functions
in computeds -- exactly the optimizer's pattern -- so only the import paths change
across the 4 consumers (store, simulation.ts, 2 test files). Golden + simulator
suites unchanged. No behavior change.
