+++ src/simulation/simulator.test.ts
@@ -3,8 +3,8 @@ import { compileScenario } from './compiler/compileScenario';
 import type { Action, ScenarioData, ScenarioTrack } from './compiler/types';
 import { simulate } from './simulator';
 import { TriggerRegistry } from './engine/TriggerRegistry';
-import { compileEndaxisScenario } from './adapters/compileEndaxisScenario';
-import { projectOptimizerResult } from './adapters/projectOptimizerResult';
+import { compileEndaxisScenario } from './compileEndaxisScenario';
+import { projectOptimizerResult } from './projection/projectOptimizerResult';
 import { projectEnemyAfflictionViz } from './projection/projectEnemyAfflictionViz';
 import { projectActionBuffs } from './projection/projectActionBuffs';
 import { createDefaultStats } from '@/simulation/defaultActorStats';
